+++ .github/release-notes.md
@@ -1,2 +1,1 @@
-* Hide spectator panel by default
-* Fixed some popups breaking when opened within a popup
+* More fixes for popups/infinite lists